# Optimize the sanitize tool