--- a/pkg/dispatcher/testdata/TestLoadConfig/good_config_with_build_farm_with_jobs.yaml
+++ b/pkg/dispatcher/testdata/TestLoadConfig/good_config_with_build_farm_with_jobs.yaml
@@ -24,7 +24,7 @@ groups:
 buildFarm:
   aws:
     build01:
-      paths:
-        - ".*some-build-farm-presubmits.yaml$"
+      filenames:
+        - "some-build-farm-presubmits.yaml"
   gcp: 
     build02: {}