Follow up https://coreos.slack.com/archives/GB7NB0CUC/p1623442582036900
TL;TR

The sanitize tool is slow because for each file (we have lots of them), we have to match the regex under `build_farm` in the config file. We do not need the regex match feature at all there. So we replace them with `sets.String.has()`.

This is a breaking change. We need a PR on the config file in release repo too.

Before the PR,
```console
time sanitize-prow-jobs --prow-jobs-dir ./ci-operator/jobs --config-path ./core-services/sanitize-prow-jobs/_config.yaml
sanitize-prow-jobs --prow-jobs-dir ./ci-operator/jobs --config-path   856.36s user 10.49s system 964% cpu 1:29.89 total
```

After,
```console
$ time sanitize-prow-jobs --prow-jobs-dir ./ci-operator/jobs --config-path ./core-services/sanitize-prow-jobs/_config.yaml
sanitize-prow-jobs --prow-jobs-dir ./ci-operator/jobs --config-path   31.00s user 4.33s system 878% cpu 4.024 total
```

/hold

/cc @stevekuznetsov